adding axios, adding missing styles, reorganizing folders
