+++ src/hooks/useGetCharacters.tsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { Character } from "../interfaces/InCharacter";
 import { Info } from "../interfaces/InInfo";
+import axios from 'axios';
 
 export const useGetCharacters = (url: string) => {
   const [characters, setCharacters] = useState<Character[]>([]);
@@ -8,17 +9,13 @@ export const useGetCharacters = (url: string) => {
   const [loading, setLoading] = useState<boolean>(true);
   const [error, setError] = useState<string | null>(null);
 
-  const loadApi = async () => {
+ const loadApi = async () => {
     try {
       setLoading(true);
-      const response = await fetch(url);
-      if (!response.ok) {
-        throw new Error(`HTTP error! Status: ${response.status}`);
-      }
-      const data = await response.json();
-      setCharacters(data.results);
-      setInfo(data.info);
-      console.log(data.info);
+      const response = await axios.get(url);
+      setCharacters(response.data.results);
+      setInfo(response.data.info);
+      console.log(response.data.info);
     } catch (error) {
       console.error("Error fetching Data: ", error);
     } finally {
